+++ p8_integration_tests/quick_test/test_array_connector/test_array_connector.py
@@ -17,7 +17,7 @@ import matplotlib.pyplot as plt
 import numpy
 from pyNN.utility.plotting import Figure, Panel
 import spynnaker8 as p
-from spynnaker8.utilities import neo_convertor
+from spynnaker.pyNN.utilities import neo_convertor
 from p8_integration_tests.base_test_case import BaseTestCase
