+++ p8_integration_tests/quick_test/test_change_neuron_parameters_between_runs/test_set_toffset.py
@@ -16,7 +16,7 @@
 import numpy
 import spynnaker.plot_utils as plot_utils
 import spynnaker8 as p
-from spynnaker8.utilities import neo_convertor
+from spynnaker.pyNN.utilities import neo_convertor
 from p8_integration_tests.base_test_case import BaseTestCase
