+++ p8_integration_tests/quick_test/test_grid_based_connectors/test_small_world_connector.py
@@ -18,7 +18,7 @@ from p8_integration_tests.base_test_case import BaseTestCase
 from pyNN.utility.plotting import Figure, Panel
 from pyNN.random import NumpyRNG
 import matplotlib.pyplot as plt
-from spynnaker8.utilities import neo_convertor
+from spynnaker.pyNN.utilities import neo_convertor
 
 
 def do_run(plot):
