+++ p8_integration_tests/quick_test/test_various/test_bigMultiProcessorSpikeSourcePrint.py
@@ -17,7 +17,7 @@
 
 import spynnaker.plot_utils as plot_utils
 import spynnaker8 as p
-from spynnaker8.utilities import neo_convertor
+from spynnaker.pyNN.utilities import neo_convertor
 from p8_integration_tests.base_test_case import BaseTestCase
