+++ p8_integration_tests/quick_test/test_various/test_synfire_Izhikevich.py
@@ -21,7 +21,7 @@ Synfirechain-like example
 import spynnaker.plot_utils as plot_utils
 import spynnaker.spike_checker as spike_checker
 import spynnaker8 as p
-from spynnaker8.utilities import neo_convertor
+from spynnaker.pyNN.utilities import neo_convertor
 from p8_integration_tests.base_test_case import BaseTestCase
