+++ p8_integration_tests/test_from_file_connector/test_from_file_connector_large.py
@@ -17,7 +17,7 @@ import os
 import matplotlib.pyplot as plt
 from pyNN.utility.plotting import Figure, Panel
 import spynnaker8 as p
-from spynnaker8.utilities import neo_convertor
+from spynnaker.pyNN.utilities import neo_convertor
 from p8_integration_tests.base_test_case import BaseTestCase
