+++ p8_integration_tests/x_test/test_0_1_time_steps/test_gsyn_allow_violate/test_synfire_0dot1_timestep_test_print_gsyn.py
@@ -24,7 +24,7 @@ from neo.io import PickleIO
 from p8_integration_tests.base_test_case import BaseTestCase
 from p8_integration_tests.scripts.synfire_run import SynfireRunner
 import spynnaker.spike_checker as spike_checker
-from spynnaker8.utilities import neo_convertor
+from spynnaker.pyNN.utilities import neo_convertor
 from spynnaker.spynnaker_plotting import SpynnakerPanel
 from spinnman.exceptions import SpinnmanTimeoutException
 from unittest import SkipTest
