+++ p8_integration_tests/x_test/test_various/spike_record_bug2.py
@@ -17,7 +17,7 @@
 
 import spynnaker.plot_utils as plot_utils
 import spynnaker8 as p
-from spynnaker8.utilities import neo_convertor
+from spynnaker.pyNN.utilities import neo_convertor
 from p8_integration_tests.base_test_case import BaseTestCase
 
 # Shows https://github.com/SpiNNakerManchester/sPyNNaker/issues/335
