+++ p8_integration_tests/x_test/test_various/test_many_spikes.py
@@ -21,7 +21,7 @@ Synfirechain-like example
 from unittest import SkipTest
 import spynnaker.plot_utils as plot_utils
 import spynnaker8 as p
-from spynnaker8.utilities import neo_convertor
+from spynnaker.pyNN.utilities import neo_convertor
 from p8_integration_tests.base_test_case import BaseTestCase
