+++ p8_integration_tests/x_test/test_various/test_mwh_population_synfire.py
@@ -22,7 +22,7 @@ from unittest import SkipTest
 from spinnman.exceptions import SpinnmanTimeoutException
 import spynnaker.plot_utils as plot_utils
 import spynnaker8 as p
-from spynnaker8.utilities import neo_convertor
+from spynnaker.pyNN.utilities import neo_convertor
 from p8_integration_tests.base_test_case import BaseTestCase
