+++ p8_integration_tests/x_test/test_various/test_poisson.py
@@ -15,7 +15,7 @@
 
 import spynnaker.plot_utils as plot_utils
 import spynnaker8 as sim
-from spynnaker8.utilities import neo_convertor
+from spynnaker.pyNN.utilities import neo_convertor
 from p8_integration_tests.base_test_case import BaseTestCase
