+++ p8_integration_tests/x_test/test_various/test_spike_io_multi_run.py
@@ -19,7 +19,7 @@ from threading import Condition
 import time
 from spinn_front_end_common.utilities.constants import NOTIFY_PORT
 import spynnaker8 as Frontend
-from spynnaker8.utilities import neo_convertor
+from spynnaker.pyNN.utilities import neo_convertor
 from p8_integration_tests.base_test_case import BaseTestCase
 
 # Create a condition to avoid overlapping prints
